added globalsleep var

+++ twitterBot.py
@@ -10,8 +10,12 @@ from random import shuffle
 import telegram_send
 from datetime import datetime
 
+# Datetime for Prints
 now = datetime.now()
 date_now = now.strftime('%Y-%m-%d %H:%M:%S')
+
+# Global Vriables
+globalsleep = 120
 
 # gets all of our data from the config file.
 with open('config.json', 'r') as config_file:
@@ -192,9 +196,11 @@ def follow_keyword(followers, following, total_followed, whitelisted_users, blac
              print(date_now)
              print('Starting to follow users who tweeted \'{}\''.format(i))
              telegram_send.send(messages=['>>>Twitter: Starting to follow users who tweeted \'{}\''.format(i)])
+                print('>>Started with Followers: ' + str(len(api.followers_ids(screen_name))))
+                print('>>Started with Following: ' + str(len(api.friends_ids(screen_name))))
              for i in range(0, len(searched_screen_names) - 1):
                 try:
-                   if int(total_followed) <= 10:
+                   if total_followed <= 10:
                     # follows the user.
                     api.create_friendship(searched_screen_names[i])
                     total_followed += 1
@@ -207,13 +213,12 @@ def follow_keyword(followers, following, total_followed, whitelisted_users, blac
                         telegram_send.send(messages=['>>>Twitter:' 'Followers: ' + str(len(api.followers_ids(screen_name)))])
                         telegram_send.send(messages=['>>>Twitter:' 'Following: ' + str(len(api.friends_ids(screen_name)))])
                     print(date_now)
-                    print('Followed user. Sleeping 120 seconds. Count: ', str(total_followed))
+                    print('Followed user. Sleeping ', globalsleep, ' seconds. Count: ', str(total_followed))
                     #telegram_send.send(messages=['>>>Twitter: ' 'Followed user. Sleeping 60 seconds.'])
-                    sleep(120)
+                    sleep(globalsleep)
                    else:
                     #print(date_now)
                     #print('countn = ' + str(total_unfollowed))
-                    while True:
                               print(date_now)
                               print('Starting to unfollow users...')
                               telegram_send.send(messages=['>>>Twitter: ' 'Starting to unfollow users...'])
@@ -221,10 +226,11 @@ def follow_keyword(followers, following, total_followed, whitelisted_users, blac
                               non_mutuals = set(following) - set(followers) - set(whitelisted_users)
                               for f in non_mutuals:
                                   try:
+                                    if total_unfollowed <= 10:
                                       # unfollows non follower.
                                       api.destroy_friendship(f)
                                       total_unfollowed += 1
-                                      total_followed -=1
+                                      total_followed -= 1
                                       if total_unfollowed % 10 == 0:
                                           print(date_now)
                                           print(str(total_unfollowed) + ' unfollowed so far.')
@@ -233,21 +239,21 @@ def follow_keyword(followers, following, total_followed, whitelisted_users, blac
                                           telegram_send.send(messages=['>>>Twitter: ' (str(total_unfollowed) + ' unfollowed so far.')])
                                           telegram_send.send(messages=['>>>Twitter: ' '>>Followers: ' + str(len(api.followers_ids(screen_name)))])
                                           telegram_send.send(messages=['>>>Twitter: ' '>>Following: ' + str(len(api.friends_ids(screen_name)))])
-                                      elif int(total_followed) <= 2:
+                                      elif total_followed <= 2:
                                           break
                                       print(date_now)
-                                      print('Unfollowed user. Sleeping 120 seconds. Count: ', str(total_unfollowed))
+                                      print('Unfollowed user. Sleeping ', globalsleep ,' seconds. Count: ', str(total_unfollowed))
                                       #telegram_send.send(messages=['>>>Twitter: ' 'Unfollowed user. Sleeping 60 seconds.'])
-                                      sleep(120)
+                                      sleep(globalsleep)
                                   except (tweepy.RateLimitError, tweepy.TweepError) as e:
                                       error_handling(e)
                               print(date_now)
-                              print(total_unfollowed)
+                              print('Total Unfollowed start: ', total_unfollowed)
 
                 except (tweepy.RateLimitError, tweepy.TweepError) as e:
                     error_handling(e)
              print(date_now)
-             print(total_followed)
+             print('Total Followed start: ', total_followed)
 
 
 # function to follow users who retweeted a tweet.
@@ -307,7 +313,7 @@ def unfollow_back(followers, following, total_followed, whitelisted_users, black
                 print(str(total_followed) + ' unfollowed so far.')
             print(date_now)
             print('Unfollowed user. Sleeping 120 seconds.')
-            sleep(120)
+            sleep(globalsleep)
         except (tweepy.RateLimitError, tweepy.TweepError) as e:
             error_handling(e)
     print(date_now)
